+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import biometric_authorization
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  BiometricAuthorizationPlugin.register(with: registry.registrar(forPlugin: "BiometricAuthorizationPlugin"))
 }